enhance: split metastore catalog interfaces into per-coordinator files
Pure reorganization: internal/metastore/catalog.go held all five *Catalog
interfaces in one file, so every coordinator's upcoming 'move my meta into
pkg/v3' PR would edit that same file and conflict with the others. Split it into
one file per coordinator (rootcoord / datacoord / querycoord / streamingcoord /
streamingnode) so each coordinator's PR touches only its own file and the moves
can land in parallel without conflicts. No interface, signature, or behaviour
change.
